Split up Booth component draggable DOM function
