+++ client/src/index.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, {useEffect} from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter as Router } from 'react-router-dom'
 import AppState from './providers/AppState'
@@ -9,7 +9,9 @@ import BoothMap from './BoothMap/'
 import './index.css'
 
 function Index() {
-
+    useEffect(() => {
+        console.clear()
+    })
     return (
         <AppState>
             <CurrentUser>
